+++ src/StardewModdingAPI/Framework/Models/ManifestDependency.cs
@@ -9,15 +9,22 @@
         /// <summary>The unique mod ID to require.</summary>
         public string UniqueID { get; set; }
 
+        /// <summary>The minimum required version (if any).</summary>
+        public ISemanticVersion MinimumVersion { get; set; }
+
 
         /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="uniqueID">The unique mod ID to require.</param>
-        public ManifestDependency(string uniqueID)
+        /// <param name="minimumVersion">The minimum required version (if any).</param>
+        public ManifestDependency(string uniqueID, string minimumVersion)
         {
             this.UniqueID = uniqueID;
+            this.MinimumVersion = !string.IsNullOrWhiteSpace(minimumVersion)
+                ? new SemanticVersion(minimumVersion)
+                : null;
         }
     }
 }